moving biolink prefix mapping to bl_utils, removing some obsolete workarounds for biolink prefix / IRI mappings
